+++ koin-projects/examples/android-samples/src/main/java/org/koin/sample/android/mvvm/MVVMActivity.kt
@@ -6,6 +6,7 @@ import kotlinx.android.synthetic.main.mvvm_activity.*
 import org.junit.Assert.*
 import org.koin.android.ext.android.getKoin
 import org.koin.android.scope.lifecycleScope
+import org.koin.android.viewmodel.ViewModelOwner.Companion.from
 import org.koin.android.viewmodel.ext.android.getViewModel
 import org.koin.android.viewmodel.ext.android.viewModel
 import org.koin.android.viewmodel.scope.viewModel
@@ -26,8 +27,8 @@ class MVVMActivity : AppCompatActivity() {
     val vm1: SimpleViewModel by viewModel(named("vm1")) { parametersOf("vm1") }
     val vm2: SimpleViewModel by viewModel(named("vm2")) { parametersOf("vm2") }
 
-    val scopeVm: ExtSimpleViewModel by lifecycleScope.viewModel(this)
-    val extScopeVm: ExtSimpleViewModel by lifecycleScope.viewModel(this, named("ext"))
+    val scopeVm: ExtSimpleViewModel by lifecycleScope.viewModel(owner = { from(this) })
+    val extScopeVm: ExtSimpleViewModel by lifecycleScope.viewModel(named("ext"), owner = { from(this) })
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -45,8 +46,8 @@ class MVVMActivity : AppCompatActivity() {
         setContentView(R.layout.mvvm_activity)
 
         supportFragmentManager.beginTransaction()
-                .replace(R.id.mvvm_frame, MVVMFragment())
-                .commit()
+            .replace(R.id.mvvm_frame, MVVMFragment())
+            .commit()
 
         getKoin().setProperty("session_id", lifecycleScope.get<Session>().id)