+++ koin-projects/koin-android-viewmodel/src/main/java/org/koin/android/viewmodel/ViewModelResolver.kt
@@ -2,6 +2,7 @@ package org.koin.android.viewmodel
 
 import android.arch.lifecycle.ViewModel
 import android.arch.lifecycle.ViewModelProvider
+import org.koin.android.viewmodel.factory.defaultViewModelFactory
 import org.koin.core.qualifier.Qualifier
 import org.koin.core.scope.Scope
 
@@ -11,9 +12,9 @@ internal fun <T : ViewModel> ViewModelProvider.resolveInstance(viewModelParamete
 }
 
 internal fun <T : ViewModel> ViewModelProvider.get(
-        viewModelParameters: ViewModelParameter<T>,
-        qualifier: Qualifier?,
-        javaClass: Class<T>
+    viewModelParameters: ViewModelParameter<T>,
+    qualifier: Qualifier?,
+    javaClass: Class<T>
 ): T {
     return if (viewModelParameters.qualifier != null) {
         get(qualifier.toString(), javaClass)
@@ -23,7 +24,14 @@ internal fun <T : ViewModel> ViewModelProvider.get(
 }
 
 internal fun <T : ViewModel> Scope.createViewModelProvider(
-        viewModelParameters: ViewModelParameter<T>
+    viewModelParameters: ViewModelParameter<T>
 ): ViewModelProvider {
-    return ViewModelProvider(viewModelParameters.viewModelStore, defaultViewModelFactory(viewModelParameters))
+    return ViewModelProvider(
+        viewModelParameters.viewModelStore, pickFactory(viewModelParameters)
+    )
+}
+
+private fun <T : ViewModel> Scope.pickFactory(
+    viewModelParameters: ViewModelParameter<T>): ViewModelProvider.Factory {
+    return defaultViewModelFactory(viewModelParameters)
 }