+++ koin-projects/koin-android-viewmodel/src/main/java/org/koin/android/viewmodel/ext/android/SharedViewModelExt.kt
@@ -16,9 +16,10 @@
 package org.koin.android.viewmodel.ext.android
 
 import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelStoreOwner
-import android.content.ComponentCallbacks
+import android.support.v4.app.Fragment
 import org.koin.android.ext.android.getKoin
+import org.koin.android.viewmodel.ViewModelOwner.Companion.from
+import org.koin.android.viewmodel.ViewModelOwnerDefinition
 import org.koin.android.viewmodel.koin.getViewModel
 import org.koin.core.parameter.ParametersDefinition
 import org.koin.core.qualifier.Qualifier
@@ -29,41 +30,38 @@ import kotlin.reflect.KClass
  *
  * @author Arnaud Giuliani
  */
-
-/**
- * LifecycleOwner extensions to help for ViewModel
- *
- * @author Arnaud Giuliani
- */
-
-fun <T : ViewModel> ViewModelStoreOwner.viewModel(
-    clazz: KClass<T>,
+inline fun <reified T : ViewModel> Fragment.sharedViewModel(
     qualifier: Qualifier? = null,
-    parameters: ParametersDefinition? = null
-): Lazy<T> {
-    return lazy(LazyThreadSafetyMode.NONE) { getViewModel(clazz, qualifier, parameters) }
-}
-
-inline fun <reified T : ViewModel> ViewModelStoreOwner.viewModel(
-    qualifier: Qualifier? = null,
+    noinline owner: ViewModelOwnerDefinition = { from(requireActivity()) },
     noinline parameters: ParametersDefinition? = null
 ): Lazy<T> {
-    return lazy(LazyThreadSafetyMode.NONE) { getViewModel<T>(qualifier, parameters) }
+    return lazy(LazyThreadSafetyMode.NONE) {
+        getViewModel<T>(qualifier, owner, parameters)
+    }
 }
 
-inline fun <reified T : ViewModel> ViewModelStoreOwner.getViewModel(
+fun <T : ViewModel> Fragment.sharedViewModel(
     qualifier: Qualifier? = null,
+    owner: ViewModelOwnerDefinition = { from(requireActivity()) },
+    clazz: KClass<T>,
+    parameters: ParametersDefinition? = null
+): Lazy<T> {
+    return lazy(LazyThreadSafetyMode.NONE) { getViewModel(qualifier, owner, clazz, parameters) }
+}
+
+inline fun <reified T : ViewModel> Fragment.getSharedViewModel(
+    qualifier: Qualifier? = null,
+    noinline owner: ViewModelOwnerDefinition = { from(requireActivity()) },
     noinline parameters: ParametersDefinition? = null
 ): T {
-    return getViewModel(T::class, qualifier, parameters)
+    return getViewModel(qualifier, owner, parameters)
 }
 
-fun <T : ViewModel> ViewModelStoreOwner.getViewModel(
-    clazz: KClass<T>,
+fun <T : ViewModel> Fragment.getSharedViewModel(
     qualifier: Qualifier? = null,
+    owner: ViewModelOwnerDefinition = { from(requireActivity()) },
+    clazz: KClass<T>,
     parameters: ParametersDefinition? = null
 ): T {
-    return getKoin().getViewModel(this, clazz, qualifier, parameters)
+    return getKoin().getViewModel(qualifier, owner, clazz, parameters)
 }
-
-private fun ViewModelStoreOwner.getKoin() = (this as ComponentCallbacks).getKoin()