+++ koin-projects/koin-android-viewmodel/src/main/java/org/koin/android/viewmodel/koin/KoinExt.kt
@@ -1,7 +1,7 @@
 package org.koin.android.viewmodel.koin
 
 import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelStoreOwner
+import org.koin.android.viewmodel.ViewModelOwnerDefinition
 import org.koin.android.viewmodel.ViewModelParameter
 import org.koin.android.viewmodel.scope.getViewModel
 import org.koin.core.Koin
@@ -10,28 +10,28 @@ import org.koin.core.qualifier.Qualifier
 import kotlin.reflect.KClass
 
 inline fun <reified T : ViewModel> Koin.viewModel(
-    owner: ViewModelStoreOwner,
     qualifier: Qualifier? = null,
+    noinline owner: ViewModelOwnerDefinition,
     noinline parameters: ParametersDefinition? = null
 ): Lazy<T> {
-    return lazy(LazyThreadSafetyMode.NONE) { getViewModel<T>(owner, qualifier, parameters) }
+    return lazy(LazyThreadSafetyMode.NONE) { getViewModel<T>(qualifier, owner, parameters) }
 }
 
 inline fun <reified T : ViewModel> Koin.getViewModel(
-    owner: ViewModelStoreOwner,
     qualifier: Qualifier? = null,
+    noinline owner: ViewModelOwnerDefinition,
     noinline parameters: ParametersDefinition? = null
 ): T {
-    return getViewModel(owner, T::class, qualifier, parameters)
+    return getViewModel(qualifier, owner, T::class, parameters)
 }
 
 fun <T : ViewModel> Koin.getViewModel(
-    owner: ViewModelStoreOwner,
-    clazz: KClass<T>,
     qualifier: Qualifier? = null,
+    owner: ViewModelOwnerDefinition,
+    clazz: KClass<T>,
     parameters: ParametersDefinition? = null
 ): T {
-    return _scopeRegistry.rootScope.getViewModel(owner, clazz, qualifier, parameters)
+    return _scopeRegistry.rootScope.getViewModel(qualifier, owner, clazz, parameters)
 }
 
 fun <T : ViewModel> Koin.getViewModel(viewModelParameters: ViewModelParameter<T>): T {